requirements.txt
# 核心依赖
torch>=1.8.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=1.0.0

# 数据处理
matplotlib>=3.3.0
seaborn>=0.11.0

# 配置管理
pyyaml>=5.4.0
hydra-core>=1.1.0

# 日志和监控
tqdm>=4.62.0
wandb>=0.12.0

# 异常检测
anomalib>=2.0.0
open_clip_torch>=2.0.0

# 隐私保护
opacus>=1.0.0

# 其他工具
python-dotenv>=0.19.0
